requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-tailwind==0.5.0
Django==3.2.18
django-appconf==1.0.5
django-apscheduler==0.6.2
django-crispy-forms==2.0
django-imagekit==4.1.0
django-mathfilters==1.0.0
django-prometheus==2.3.1
django-taggit==4.0.0
idna==3.4
pilkit==2.0
Pillow==9.5.0
prometheus-client==0.17.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==4.2
urllib3==2.0.3
